+++ serlib/ser_sorts.ml
@@ -17,21 +17,22 @@ module Univ = Ser_univ
 
 type family =
   [%import: Sorts.family]
-  [@@deriving sexp]
+  [@@deriving sexp,yojson]
 
 type _t =
   | SProp
   | Prop
   | Set
   | Type of Univ.Universe.t
-  [@@deriving of_sexp]
+  [@@deriving of_sexp,yojson]
 
 type t =
   [%import: Sorts.t]
-  [@@deriving sexp_of]
+  [@@deriving sexp_of,to_yojson]
 
 let t_of_sexp x = Obj.magic (_t_of_sexp x)
+let of_yojson json = Ppx_deriving_yojson_runtime.(_t_of_yojson json >|= Obj.magic)
 
 type relevance =
   [%import: Sorts.relevance]
-  [@@deriving sexp]
+  [@@deriving sexp,yojson]