+++ serlib/ser_univ.mli
@@ -15,8 +15,8 @@
 
 open Sexplib
 
-module Level : SerType.S with type t = Univ.Level.t
-module Universe : SerType.S with type t = Univ.Universe.t
+module Level : SerType.SJ with type t = Univ.Level.t
+module Universe : SerType.SJ with type t = Univ.Universe.t
 
 module Variance : SerType.S with type t = Univ.Variance.t
 module Instance : SerType.S with type t = Univ.Instance.t
@@ -31,7 +31,7 @@ type univ_constraint = Univ.univ_constraint
 val univ_constraint_of_sexp : Sexp.t -> univ_constraint
 val sexp_of_univ_constraint : univ_constraint -> Sexp.t
 
-module Constraint : SerType.S with type t = Univ.Constraint.t
+module Constraint : SerType.SJ with type t = Univ.Constraint.t
 module UContext : SerType.S with type t = Univ.UContext.t
 
 module AUContext : SerType.S with type t = Univ.AUContext.t
@@ -51,6 +51,9 @@ type 'a puniverses = 'a * Instance.t
 
 val puniverses_of_sexp : (Sexp.t -> 'a) -> Sexp.t -> 'a puniverses
 val sexp_of_puniverses : ('a -> Sexp.t) -> 'a puniverses -> Sexp.t
+
+val puniverses_of_yojson : (Yojson.Safe.t -> ('a, string) Result.result) -> Yojson.Safe.t -> ('a puniverses, string) Result.result
+val puniverses_to_yojson : ('a -> Yojson.Safe.t) -> 'a puniverses -> Yojson.Safe.t
 
 type explanation = Univ.explanation